Moves emeritus to their own page
