+++ app/controllers/teams/index.js
@@ -4,7 +4,7 @@ import { cached } from '@glimmer/tracking';
 const inTeam = (team) => (teamMember) =>
   (teamMember.teams ?? []).includes(team);
 
-export default class TeamsController extends Controller {
+export default class TeamsIndexController extends Controller {
   @cached
   get sortedTeamMembers() {
     const teamMembers = this.model?.slice() ?? [];
@@ -12,10 +12,6 @@ export default class TeamsController extends Controller {
     return teamMembers.sort((teamMember1, teamMember2) => {
       return teamMember1.added - teamMember2.added;
     });
-  }
-
-  get alumniTeamMembers() {
-    return this.sortedTeamMembers.filter(inTeam('alumni'));
   }
 
   get coreToolingTeamMembers() {
