+++ app/routes/teams/emeritus.js
@@ -1,0 +1,3 @@
+import Route from '@ember/routing/route';
+
+export default class TeamsEmeritusRoute extends Route {}
